[SPARK-51118][TESTS][FOLLOWUP] Move test_udf_with_udt to BaseUDFTestsMixin
Move `test_udf_with_udt` to `BaseUDFTestsMixin`.

These tests should be shared properly.

No.

The moved tests should pass.

No.

Closes #50392 from ueshin/issues/SPARK-51118/test.

Authored-by: Takuya Ueshin <ueshin@databricks.com>
